Reorganize `sdd_builder` into submodules
Things are getting unwieldly!
